# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VulnScanner Enterprise API",
    "description": "Microservicio web para automatización de auditorías de seguridad, Reportes Ejecutivos PDF, Auto-PR GitHub DevSecOps, IAST/RASP, Grafos de Ataque, OpenAPI y Dashboard SOC en tiempo real.",
    "version": "2.4.0"
  },
  "paths": {
...
    "/download": {
      "get": {
        "summary": "Download Report",
        "description": "Permite la descarga segura de reportes generados (HTML, JSON, SARIF, PDF).",
        "operationId": "download_report_download_get",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/scan": {
      "post": {
        "summary": "Start Scan",
        "description": "Inicia un escaneo web en segundo plano y responde inmediatamente con un ID de tarea.",
        "operationId": "start_scan_scan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Start Scan Scan Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/scan/{task_id}": {
      "get": {
        "summary": "Get Scan Status",
        "description": "Retorna el estado actual de una tarea de escaneo específica y sus resultados si terminó.",
        "operationId": "get_scan_status_scan__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Scan Status Scan  Task Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ScanRequest": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "stealth": {
            "type": "boolean",
            "title": "Stealth",
            "default": true
          },
          "delay": {
            "type": "number",
            "title": "Delay",
            "default": 0.0
          },
          "crawl": {
            "type": "integer",
            "title": "Crawl",
            "default": 1
          },
          "subdomains": {
            "type": "boolean",
            "title": "Subdomains",
            "default": false
          },
          "cookie": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cookie"
          },
          "auth": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auth"
          },
          "webhook_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webhook Url"
          },
          "passive": {
            "type": "boolean",
            "title": "Passive",
            "default": false
          },
          "enable_oast": {
            "type": "boolean",
            "title": "Enable Oast",
            "default": true
          },
          "profile": {
            "type": "string",
            "title": "Profile",
            "default": "normal"
          },
          "allow_private": {
            "type": "boolean",
            "title": "Allow Private",
            "default": false
          },
          "har_file": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Har File"
          },
          "headless_crawl": {
            "type": "boolean",
            "title": "Headless Crawl",
            "default": false
          },
          "headless_login": {
            "type": "boolean",
            "title": "Headless Login",
            "default": false
          },
          "openapi_spec": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Openapi Spec"
          },
          "use_async_engine": {
            "type": "boolean",
            "title": "Use Async Engine",
            "default": false
          },
          "no_waf_detect": {
            "type": "boolean",
            "title": "No Waf Detect",
            "default": false
          },
          "iast_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iast Url"
          },
          "enable_attack_chain": {
            "type": "boolean",
            "title": "Enable Attack Chain",
            "default": true
          },
          "generate_pdf": {
            "type": "boolean",
            "title": "Generate Pdf",
            "default": true
          },
          "auto_pr": {
            "type": "boolean",
            "title": "Auto Pr",
            "default": false
          },
          "github_repo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Repo"
          },
          "github_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Token"
          },
          "base_branch": {
            "type": "string",
            "title": "Base Branch",
            "default": "main"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "ScanRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}